+++ src/module/integration.js
@@ -4,3 +4,4 @@ export const MODULE_DAE = "dae";
 export const MODULE_DSN = "dice-so-nice";
 export const MODULE_LIB = "lib-wrapper";
 export const MODULE_MIDI = "midi-qol";
+export const MODULE_AC5E = "automated-conditions-5e";